split up plan and load tests
